+++ src/main/java/com/epam/informationhandling/logic/expressioncalculation/TerminalDivideExpression.java
@@ -3,8 +3,8 @@ package com.epam.informationhandling.logic.expressioncalculation;
 public class TerminalDivideExpression implements MathExpression {
     @Override
     public void interpret(Context context) {
-        Integer secondMember = context.pop();
-        Integer firstMember = context.pop();
+        Double secondMember = context.pop();
+        Double firstMember = context.pop();
         context.push(firstMember / secondMember);
     }
 }
